Ctrl-Option-F5 and Ctrl-Option-Return instead of Ctrl-F5 and Ctrl-Return
This code changes the "hotkeys" for mouse grabbing and full-screen
toggle from Ctrl-F5 and Ctrl-Return to Ctrl-Option-F5 and
Ctrl-Option-Return to avoid conflicts with OS X keystrokes.

The code that refers to meta keys makes it possible to use Cmd (or Win)
in addition to Ctrl and Alt in hotkeys.

A similar change should probably be made to Ctrl-F1 hotkey for
capturing a floppy, but I haven't done so in this set of changes.

This is a temporary solution.